+++ src/Component/develop.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable max-lines */
 // noinspection ES6UnusedImports,JSUnusedLocalSymbols
 
-import React, { useCallback, useRef, useState } from 'react';
+import React, {EventHandler, MouseEvent, MouseEventHandler, useCallback, useRef, useState} from 'react';
 import {
   Button,
   ButtonProps,
@@ -31,39 +31,36 @@ import {
 import {
   AccountCircle as AccountCircleIcon,
   Block,
-  CheckBox,
   ContactPhone as ContactPhoneIcon,
-  Error,
   Refresh,
   Visibility,
-  VisibilityOff
 } from '@material-ui/icons';
-import { useTheme } from '@material-ui/core/styles';
-import { getCurrency } from 'Utilities/Format/Currency';
-import { useDispatch } from 'react-redux';
+import {useTheme} from '@material-ui/core/styles';
+import {getCurrency} from 'Utilities/Format/Currency';
+import {useDispatch} from 'react-redux';
 import * as Action from 'Store/Action';
-import { VirtualizedTable } from 'Component/Table';
-import { Index, TableCellProps } from 'react-virtualized';
-import TextFieldClearButton, { TextFieldClearButtonProps } from 'Component/TextFieldClearButton';
-import { ContentEllipseMode } from 'Component/Table/TableCell/TableCellDefault';
-import { TableCellDateTime, TableCellDateTimeRange, TableCellDefault, TableCellProduct } from 'Component/Table/TableCell';
-import { ButtonProgress } from 'Component/Button';
-import { AppCardContent, AppCardHeader } from 'Component/Card';
-import { TableRowLoading } from 'Component/Table/TableRow';
+import {VirtualizedTable} from 'Component/Table';
+import {Index, TableCellProps} from 'react-virtualized';
+import TextFieldClearButton, {TextFieldClearButtonProps} from 'Component/TextFieldClearButton';
+import {ContentEllipseMode} from 'Component/Table/TableCell/TableCellDefault';
+import {TableCellDateTime, TableCellDateTimeRange, TableCellDefault, TableCellProduct} from 'Component/Table/TableCell';
+import {ButtonProgress} from 'Component/Button';
+import {AppCardContent, AppCardHeader} from 'Component/Card';
+import {TableRowLoading} from 'Component/Table/TableRow';
 import SearchIcon from '@material-ui/icons/Search';
-import { CustomerCard, FilterReset } from 'Component/Icon';
-import { ColumnData } from 'Component/Table/VirtualizedTable';
-import { DateTimeRangeCellProps } from 'Types/Table/DateTimeRangeCellProps';
+import {CustomerCard, FilterReset} from 'Component/Icon';
+import {ColumnData} from 'Component/Table/VirtualizedTable';
+import {DateTimeRangeCellProps} from 'Types/Table/DateTimeRangeCellProps';
 import WarningIcon from '@material-ui/icons/Warning';
 import TableCellSelect from 'Component/Table/TableCell/TableCellSelect';
 import TableHeadSelect from 'Component/Table/TableHead/TableHeadSelect';
-import { OptionsObject, VariantType } from 'notistack';
-import { Notification } from 'Types/Notification';
+import {OptionsObject, VariantType} from 'notistack';
+import {Notification} from 'Types/Notification';
 import Box from '@material-ui/core/Box';
-import { NativeSelectProps } from '@material-ui/core/NativeSelect/NativeSelect';
+import {NativeSelectProps} from '@material-ui/core/NativeSelect/NativeSelect';
 import DatePicker from 'Component/DatePicker';
-import SelectClearButton, { SelectClearButtonProps } from 'Component/SelectClearButton';
-import { TableCellProductProps } from 'Component/Table/TableCell/TableCellProduct';
+import SelectClearButton, {Props as SelectClearButtonProps} from 'Component/SelectClearButton';
+import {TableCellProductProps} from 'Component/Table/TableCell/TableCellProduct';
 
 const tableRowHeight = 48;
 
@@ -81,7 +78,7 @@ const generateNotification = (persist = false): Notification => {
     id,
     group,
     variant,
-    notistackOptions: { persist: persist }
+    notistackOptions: {persist: persist}
   };
   switch (variant) {
     case 'error':
@@ -107,7 +104,7 @@ const generateNotification = (persist = false): Notification => {
         action: () => <IconButton><Visibility/></IconButton>
       };
     case 'success':
-      return { ...defaultProps, title: 'Changes saved' };
+      return {...defaultProps, title: 'Changes saved'};
     case 'info':
       return {
         ...defaultProps,
@@ -116,7 +113,7 @@ const generateNotification = (persist = false): Notification => {
       };
     case 'default':
     default:
-      return { ...defaultProps, variant: 'default', content: 'content loaded' };
+      return {...defaultProps, variant: 'default', content: 'content loaded'};
   }
 };
 
@@ -138,10 +135,10 @@ const sample: SampleVirtualizedTable[] = [
     end: null,
     isLiveShow: false
   }, null],
-  ['Ice cream sandwich', 237, 9.0, 37, { start: new Date(), end: new Date(), isLiveShow: false }, new Date()],
-  ['Eclair', 262, 16.0, 24, { start: new Date(2021, 9, 23, 8, 0, 15), end: new Date(), isLiveShow: false }, new Date()],
-  ['Cupcake', 305, 3.7, 67, { start: new Date(), end: new Date(), isLiveShow: false }, new Date()],
-  ['Gingerbread', 356, 16.0, 49, { start: new Date(), end: new Date(), isLiveShow: false }, new Date()]
+  ['Ice cream sandwich', 237, 9.0, 37, {start: new Date(), end: new Date(), isLiveShow: false}, new Date()],
+  ['Eclair', 262, 16.0, 24, {start: new Date(2021, 9, 23, 8, 0, 15), end: new Date(), isLiveShow: false}, new Date()],
+  ['Cupcake', 305, 3.7, 67, {start: new Date(), end: new Date(), isLiveShow: false}, new Date()],
+  ['Gingerbread', 356, 16.0, 49, {start: new Date(), end: new Date(), isLiveShow: false}, new Date()]
 ];
 
 const createDataVirtualizedTable = (id: number, dessert: string, calories: number, fat: number, carbs: number, datetimeRange: DateTimeRangeCellProps, datetime?: Date): DataVirtualizedTable => ({
@@ -208,7 +205,7 @@ const columns: ColumnData[] = [
     cellRenderer: cellProps => <TableCellDateTime cellData={cellProps.cellData} height={tableRowHeight}/>
   }
 ];
-const textFieldInputProps = { startAdornment: <InputAdornment position={'start'}><SearchIcon/></InputAdornment> };
+const textFieldInputProps = {startAdornment: <InputAdornment position={'start'}><SearchIcon/></InputAdornment>};
 
 const createProductDataVirtualizedTable = (index: number): TableCellProductProps => ({
   rowIndex: index,
@@ -231,23 +228,18 @@ for (let i = 0; i < 200; i += 1) {
   productRows.push(createProductDataVirtualizedTable(i));
 }
 
-const selectMenuItems = [
-  <MenuItem value={'test 1'} key={'menu-item-1'}>Test 1</MenuItem>,
-  <MenuItem value={'test 2'} key={'menu-item-2'}>Test 2</MenuItem>,
-  <MenuItem value={'test 3'} key={'menu-item-3'}>Test 3</MenuItem>,
-  <MenuItem value={'test 4'} key={'menu-item-4'}>Test 4</MenuItem>
-];
-
 const Develop: React.FC = () => {
   const theme = useTheme();
   const dispatch = useDispatch();
   const [testTextFieldValue, setTestTextFieldValue] = useState('');
-  const [testSelectValue, setTestSelectValue] = useState('');
+  const [singleSelectValue, setSingleSelectValue] = useState('');
+  const [multipleSelectValue, setMultipleSelectValue] = useState([]);
   const [testDatePickerValue, setTestDatePickerValue] = useState<Date>(null);
   const noRowsRenderer = useCallback(() => <TableRowLoading/>, []);
   const rowGetter = useCallback((row: Index) => rows[row.index], []);
   const productRowGetter = useCallback((row: Index) => productRows[row.index], []);
-  const handleSelectUpdate: SelectClearButtonProps['onChange'] = useCallback(event => setTestSelectValue(event === null ? '' : event.target.value as string), []);
+  const handleSingleSelectUpdate: SelectClearButtonProps['onChange'] = useCallback(value => setSingleSelectValue(value as string), []);
+  const handleMultipleSelectUpdate: SelectClearButtonProps['onChange'] = useCallback(value => setMultipleSelectValue(value as string[]), []);
   const handleTextFieldUpdate: TextFieldClearButtonProps['onChange'] = useCallback(event => setTestTextFieldValue(event === null ? '' : event.target.value), []);
   const persistNotificationsRef = useRef(null);
   const handleOnClickNotification: ButtonProps['onClick'] = useCallback(() => {
@@ -301,12 +293,79 @@ const Develop: React.FC = () => {
   })), []);
 
   return (
-    <div style={{ margin: theme.spacing(1) }}>
+    <div style={{margin: theme.spacing(1)}}>
       <Card>
         <AppCardHeader title={'Test'} titleIcon={<WarningIcon/>} onRefresh={console.log}/>
         <AppCardContent>
           Das ist der Inhalt
           <IconButton size={'small'} color={'secondary'}><FilterReset/></IconButton>
+          <Grid container spacing={2}>
+            <Grid item>
+              <DatePicker
+                label={'DatePicker'}
+                onChange={console.log}
+                value={testDatePickerValue}
+                isClearable={false}
+              />
+            </Grid>
+            <Grid item>
+              <TextFieldClearButton
+                label={'Textfield'}
+                onChange={handleTextFieldUpdate}
+                value={testTextFieldValue}
+                clearButtonSize={'small'}
+                InputProps={textFieldInputProps}
+              />
+            </Grid>
+            <Grid item>
+              <TextFieldClearButton
+                disabled
+                label={'Textfield'}
+                onChange={handleTextFieldUpdate}
+                value={testTextFieldValue}
+                clearButtonSize={'small'}
+              />
+            </Grid>
+            <Grid item xs={1}>
+              <SelectClearButton
+                fullWidth
+                label={'Single select'}
+                onChange={handleSingleSelectUpdate}
+                value={singleSelectValue}
+                clearButtonSize={'small'}
+                renderValueAsChip
+                options={[
+                  {value: 'test 1', label: 'Test 1'},
+                  {value: 'test 2', label: 'Test 2'},
+                  {value: 'test 3', label: 'Test 3'},
+                  {value: 'test 4', label: 'Test 4'},
+                  {value: 'DasIstEinSehrLangerWert', label: 'Das ist ein sehr langer Wert'},
+                  {value: 'DasIstEinSehrLangerWertMal2', label: 'Das ist ein sehr langer Wert Das ist ein sehr langer Wert'},
+                ]}
+              />
+            </Grid>
+            <Grid item xs={1}>
+              <SelectClearButton
+                fullWidth
+                label={'Multiple select'}
+                onChange={handleMultipleSelectUpdate}
+                value={multipleSelectValue}
+                clearButtonSize={'small'}
+                multiple
+                maxValuesToDisplayInInput={2}
+                renderValueAsChip
+                options={[
+                  {value: 'test 1', label: 'Test 1 Das ist ein sehr langer Wert Wert Wert Wert'},
+                  {value: 'test 2', label: 'Test 2'},
+                  {value: 'test 3', label: 'Test 3'},
+                  {value: 'test 4', label: 'Test 4'},
+                  {value: 'DasIstEinSehrLangerWert', label: 'Das ist ein sehr langer Wert'},
+                  {value: 'DasIstEinSehrLangerWertMal2', label: 'Das ist ein sehr langer Wert Das ist ein sehr langer Wert'},
+                ]}
+              />
+            </Grid>
+            <Grid item><Button onClick={() => setMultipleSelectValue(['test 1'])} color={'primary'}>Set Multiple Select Value</Button></Grid>
+          </Grid>
         </AppCardContent>
       </Card>
       {/*<div style={{ marginTop: theme.spacing(1) }}>*/}
@@ -326,40 +385,6 @@ const Develop: React.FC = () => {
       {/*</Card>*/}
       {/*</div>*/}
       <Grid container spacing={1} alignItems={'center'}>
-        <Grid item>
-          <DatePicker
-            label={'DatePicker'}
-            onChange={console.log}
-            value={testDatePickerValue}
-            isClearable={false}
-          />
-        </Grid>
-        <Grid item>
-          <TextFieldClearButton
-            label={'Textfield'}
-            onChange={handleTextFieldUpdate}
-            value={testTextFieldValue}
-            clearButtonSize={'small'}
-            InputProps={textFieldInputProps}
-          />
-        </Grid>
-        <Grid item>
-          <TextFieldClearButton
-            disabled
-            label={'Textfield'}
-            onChange={handleTextFieldUpdate}
-            value={testTextFieldValue}
-            clearButtonSize={'small'}
-          />
-        </Grid>
-        <Grid item>
-          <SelectClearButton
-            label={'Select with Clear Button'}
-            onChange={handleSelectUpdate}
-            value={testSelectValue}
-            clearButtonSize={'small'}
-          >{selectMenuItems}</SelectClearButton>
-        </Grid>
         <Grid item>
           <CustomerCard/>
         </Grid>
@@ -436,7 +461,7 @@ const Develop: React.FC = () => {
       {/*    </ListItem>*/}
       {/*  </List>*/}
       {/*</div>*/}
-      <div style={{ marginTop: theme.spacing(1) }}>
+      <div style={{marginTop: theme.spacing(1)}}>
         <Grid container>
           <Grid item>
             <Typography
@@ -473,7 +498,7 @@ const Develop: React.FC = () => {
           </Grid>
         </Grid>
       </div>
-      <div style={{ marginTop: theme.spacing(1) }}>
+      <div style={{marginTop: theme.spacing(1)}}>
         <Grid container spacing={1}>
           <Grid item xs={3}>
             <Paper>
@@ -529,10 +554,10 @@ const Develop: React.FC = () => {
         </Grid>
       </div>
 
-      <div style={{ marginTop: theme.spacing(1) }}>
+      <div style={{marginTop: theme.spacing(1)}}>
         <Card>
-          <CardContent style={{ height: 300 }}>
-            <TableContainer style={{ maxHeight: 260 }}>
+          <CardContent style={{height: 300}}>
+            <TableContainer style={{maxHeight: 260}}>
               <Table stickyHeader>
                 <TableHead>
                   <TableRow>
@@ -554,9 +579,9 @@ const Develop: React.FC = () => {
           </CardContent>
         </Card>
       </div>
-      <div style={{ marginTop: theme.spacing(1) }}>
+      <div style={{marginTop: theme.spacing(1)}}>
         <Card>
-          <CardContent style={{ height: 300 }}>
+          <CardContent style={{height: 300}}>
             <VirtualizedTable
               headerHeight={50}
               rowHeight={100}